refactor(utils): move copyTextToClipboard to superset-ui-core
The copy utility had no app-specific dependencies (just browser APIs),
so it belongs in the package layer. Moving it to
superset-ui-core/src/utils/copy.ts and re-exporting from
src/utils/copy.ts keeps all existing callers working without changes.

CodeSyntaxHighlighter now uses the shared utility instead of an
inline duplicate, resolving the DRY violation flagged in review.
